requirements.txt
annotated-types==0.7.0
anyio==4.5.2
bcrypt==4.3.0
beautifulsoup4==4.13.4
blinker==1.8.2
boto3==1.37.38
botocore==1.37.38
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
cryptography==45.0.6
deprecation==2.1.0
distro==1.9.0
exceptiongroup==1.3.0
Flask==3.0.3
Flask-Cors==5.0.0
Flask-JWT-Extended==4.6.0
gotrue==2.9.2
gunicorn==21.2.0
h11==0.16.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httpcore==1.0.9
httpx==0.27.2
hyperframe==6.0.1
idna==3.10
importlib_metadata==8.5.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.1
jmespath==1.0.1
MarkupSafe==2.1.5
newsapi-python==0.2.7
azure-cognitiveservices-speech==1.40.0
openai==1.99.3
packaging==25.0
pluggy==1.5.0
postgrest==0.16.11
praw==7.8.1
prawcore==2.4.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
pytest==8.3.5
pytest-flask==1.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
realtime==1.0.6
requests==2.32.4
s3transfer==0.11.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
storage3==0.7.7
StrEnum==0.4.15
stripe==12.4.0
supabase==2.6.0
supafunc==0.5.1
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.13.2
update-checker==0.18.0
urllib3==1.26.20
uuid==1.30
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.0.6
wordfreq
zipp==3.20.2
psutil
tenacity==8.2.3
spacy==3.7.2
lemminflect==0.2.3
jieba==0.42.1
fugashi==1.3.0
unidic-lite==1.0.8